% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_func_W_dV}
\alias{set_func_W_dV}
\title{Set the wetness index function}
\usage{
set_func_W_dV(newfunc)
}
\arguments{
\item{newfunc}{a function which computes wetness index \code{W} [unitless]
from storage deficit \code{dV} [mm],
with arguments \code{x} (here storage deficit).
If ran with newfunc=NULL, the function will be reset to the default.}
}
\value{
a function to be used as evapotranspiration reduction function.
}
\description{
Changes the default wetness index function \code{func_W_dV}
}
\examples{
x=1
}

