% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_func_beta_dV}
\alias{show_func_beta_dV}
\title{Show current evapotranspiration reduction function}
\usage{
show_func_beta_dV()
}
\value{
the function to be used as evapotranspiration reduction function.
}
\description{
Shows which evapotranspiration reduction function \code{func_beta_dV} is used.
}
\examples{
x=1
}

