% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_defaults.R
\name{show_func_dVeq_dG}
\alias{show_func_dVeq_dG}
\title{Show current equilibrium storage deficit-groundwater depth relation}
\usage{
show_func_dVeq_dG()
}
\value{
the function to be used as relation between equilibrium storage deficit and groundwater depth.
}
\description{
Shows which relation between equilibrium storage deficit and groundwater
depth \code{func_dVeq_dG} is used.
}
\examples{
x=1
}

