% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_balance.R
\name{WALRUS_balance}
\alias{WALRUS_balance}
\title{Computes the water budget.}
\usage{
WALRUS_balance(o, pars, n)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{pars}{the parameter set used for the run.}

\item{n}{the name of the run.}
}
\value{
a data frame with the water budget, both printed to screen and written to file (in the output-folder).
}
\description{
Computes the water budget over the WALRUS-run.
}
\examples{
x=1
}

