% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_figures}
\alias{WALRUS_figures}
\title{Plot WALRUS-output}
\usage{
WALRUS_figures(o, pars, n)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{pars}{the parameter set used for the run.}

\item{n}{the name of the run.}
}
\value{
a figure in R and a figure in pdf.
}
\description{
Makes several figures of the output variables from WALRUS.
}
\examples{
x=1
}

