% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_GoF}
\alias{WALRUS_GoF}
\title{Computes goodness of fit for WALRUS}
\usage{
WALRUS_GoF(o, pars, n)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{pars}{the parameter set used for the run.}

\item{n}{the name of the run.}
}
\value{
NS, logNS and meanSS
}
\description{
Computes the Nash-Sutcliffe efficiency of the discharge, Nash-Sutcliffe efficiency of the logarithm of the discharge and the mean sum of squares of the discharge
}
\examples{
x=1
}

