% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_loop_long.R
\name{WALRUS_loop_long}
\alias{WALRUS_loop_long}
\title{Use this function instead of \code{WALRUS_loop} when using very long time series.
Divides the time series in chunks, calls \code{WALRUS_loop} and pasts output together.}
\usage{
WALRUS_loop_long(pars, chunk_size = 10000, show_progress = FALSE)
}
\arguments{
\item{pars}{the parameter set used for the run.}

\item{chunk_size}{the maximum number of time steps to run in one go (default 200,000).}

\item{show_progress}{whether the progress should be printed to screen (default false).}
}
\value{
a data frame with the model output for all output time steps.
}
\description{
One of the main WALRUS-functions. Computes initial values and provides the for-loop with flexible time step.
}
\examples{
x=1
}

