% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_preprocessing}
\alias{WALRUS_preprocessing}
\title{Preprocessing}
\usage{
WALRUS_preprocessing(f, dt)
}
\arguments{
\item{f}{a dataframe with forcing data: date in YYYYmmddhh or YYYmmdd format and columns with P, ETpot, Q (at least one value), fXG (optional), fXS (optional), dG (optional)}

\item{dt}{the time step size increase (e.g. when given hourly data and want to producre daily model output, use 24.)}
}
\value{
a vector \code{output_date} with the moments on which output should be generated
}
\description{
Preprocesses the forcing data.
}
\examples{
x=1
}

