% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_preprocessing_calibration.R
\name{WALRUS_preprocessing_calibration}
\alias{WALRUS_preprocessing_calibration}
\title{Preprocessing for calibration}
\usage{
WALRUS_preprocessing_calibration()
}
\value{
a data frame with the water budget.
}
\description{
Converts discharge and/or groundwater times series to same time steps as output_steps.
}
\examples{
x=1
}

