% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_residuals.R
\name{WALRUS_residuals}
\alias{WALRUS_residuals}
\title{Plot residual plots}
\usage{
WALRUS_residuals(o, pars, n)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{pars}{the parameter set used for the run.}

\item{n}{the name of the run.}
}
\value{
a figure with residual plots
}
\description{
Plot figures with residuals from WALRUS.
}
\examples{
x=1
}

