% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_selectdates.R
\name{WALRUS_selectdates}
\alias{WALRUS_selectdates}
\title{Select dates from a forcing data frame.}
\usage{
WALRUS_selectdates(dataframe, start, end)
}
\arguments{
\item{dataframe}{a dataframe with date in YYYYmmddhh or YYYmmdd format and columns with P, ETpot, Q (at least one value), fXG (optional), fXS (optional), GWL (optional)}

\item{start}{start of the simulation perio (in YYYYmmddhh or YYYmmdd format)}

\item{end}{end of the simulation perio (in YYYYmmddhh or YYYmmdd format)}
}
\value{
a shorter data frame of the \code{dataframe} input
}
\description{
Select a shorter period from a forcing data frame.
}
\examples{
x=1
}

