% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_shiny_figures.R
\name{WALRUS_shiny_figures}
\alias{WALRUS_shiny_figures}
\title{Plot WALRUS-output for shiny}
\usage{
WALRUS_shiny_figures(o, pars)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{pars}{the parameter set used for the run.}
}
\value{
a figure in R.
}
\description{
Makes several figures of the output variables from WALRUS.
}
\examples{
x=1
}

