% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_step.R
\name{WALRUS_step}
\alias{WALRUS_step}
\title{Computes variables for the next time step.}
\usage{
WALRUS_step(pars, i, t1, t2)
}
\arguments{
\item{pars}{the parameter set used for the run.}

\item{i}{the initital states for this time step}

\item{t1}{the start of the time step (in seconds since 1970-01-01)}

\item{t2}{the end of the time step (in seconds since 1970-01-01)}
}
\value{
a vector with the model output for one time step.
}
\description{
One of the main WALRUS-functions. Computes states and fluxes for one time step.
}
\examples{
x=1
}

