% Generated by roxygen2 (4.0.1): do not edit by hand
\name{show_func_W_dV}
\alias{show_func_W_dV}
\title{Show current wetness index function}
\usage{
show_func_W_dV()
}
\value{
the function to be used as wetness index function.
}
\description{
Shows which wetness index function \code{func_W_dV} is used.
}
\examples{
x=1
}

