% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WALRUS_output_file}
\alias{WALRUS_output_file}
\title{Makes output files.}
\usage{
WALRUS_output_file(o, n)
}
\arguments{
\item{o}{the output of a WALRUS run (result of \code{WALRUS_loop}).}

\item{n}{the name of the run.}
}
\value{
a data file with all model output
}
\description{
Creates a data file with all model output.
}
\examples{
x=1
}

