% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WALRUS_preprocessing.R
\name{WALRUS_preprocessing}
\alias{WALRUS_preprocessing}
\title{Preprocessing}
\usage{
WALRUS_preprocessing(f, dt, timestamp = "start")
}
\arguments{
\item{f}{a dataframe with forcing data: date in YYYYmmddhh or YYYmmdd format and
columns with P, ETpot, Q (at least one value), fXG (optional), fXS (optional), dG (optional)}

\item{dt}{the time step size increase (e.g. when given hourly data and want to
produce daily model output, use 24.)}

\item{timestamp}{can be either "start" (default) or "end". Signifies whether the dates belong to
the start or the end of the measurement interval. Example: using hourly data, the discharge data
behind 2015010105 is the discharge sum between 5 and 6 o'clock when timestamp="start" and
the sum between 4 and 5  o'clock when timestamp="end".}
}
\value{
a vector \code{output_date} with the moments on which output should be generated
}
\description{
Preprocesses the forcing data.
}
\examples{
x=1
}

